# front_portal_fejepe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portal FEJEPE API",
    "description": "API do Portal FEJEPE — centraliza dados e cálculos estratégicos das Empresas Juniores da rede. Permite consultar EJs, comparar indicadores, visualizar KPIs da rede e gerar rankings.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rede/indicadores": {
      "get": {
        "tags": [
          "Rede"
        ],
        "summary": "Indicadores da rede",
        "description": "Retorna os KPIs consolidados da rede FEJEPE: faturamento total, faturamento colaborativo, média de CSAT, número de EJs fora do zero, distribuição por cluster, participação por comunidade e crescimento.",
        "operationId": "get_indicadores_rede_api_v1_rede_indicadores_get",
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Ano de referência",
              "default": 2026,
              "title": "Ano"
            },
            "description": "Ano de referência"
          },
          {
            "name": "mes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Mês de referência",
              "title": "Mes"
            },
            "description": "Mês de referência"
          },
          {
            "name": "cluster",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por cluster",
              "title": "Cluster"
            },
            "description": "Filtrar por cluster"
          },
          {
            "name": "comunidade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por comunidade",
              "title": "Comunidade"
            },
            "description": "Filtrar por comunidade"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndicadoresRede"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rede/ranking": {
      "get": {
        "tags": [
          "Rede"
        ],
        "summary": "Ranking de EJs",
        "description": "Gera ranking dinâmico de EJs por critério: faturamento, CSAT, projetos ou faturamento colaborativo. Pode ser filtrado por cluster e comunidade.",
        "operationId": "get_ranking_api_v1_rede_ranking_get",
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Ano de referência",
              "default": 2026,
              "title": "Ano"
            },
            "description": "Ano de referência"
          },
          {
            "name": "mes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Mês de referência",
              "title": "Mes"
            },
            "description": "Mês de referência"
          },
          {
            "name": "criterio",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(faturamento|csat|projetos|colab)$",
              "description": "Critério de ranking: faturamento, csat, projetos, colab",
              "default": "faturamento",
              "title": "Criterio"
            },
            "description": "Critério de ranking: faturamento, csat, projetos, colab"
          },
          {
            "name": "cluster",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por cluster",
              "title": "Cluster"
            },
            "description": "Filtrar por cluster"
          },
          {
            "name": "comunidade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por comunidade",
              "title": "Comunidade"
            },
            "description": "Filtrar por comunidade"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Máximo de itens no ranking",
              "default": 20,
              "title": "Limit"
            },
            "description": "Máximo de itens no ranking"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankingItem"
                  },
                  "title": "Response Get Ranking Api V1 Rede Ranking Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IndicadoresRede": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "mes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mes"
          },
          "faturamento_total": {
            "type": "number",
            "title": "Faturamento Total",
            "default": 0
          },
          "faturamento_colab_total": {
            "type": "number",
            "title": "Faturamento Colab Total",
            "default": 0
          },
          "total_projetos": {
            "type": "integer",
            "title": "Total Projetos",
            "default": 0
          },
          "media_csat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Csat"
          },
          "total_ejs": {
            "type": "integer",
            "title": "Total Ejs",
            "default": 0
          },
          "ejs_fora_do_zero": {
            "type": "integer",
            "title": "Ejs Fora Do Zero",
            "default": 0
          },
          "ejs_ritmo_minimo": {
            "type": "integer",
            "title": "Ejs Ritmo Minimo",
            "default": 0
          },
          "ejs_ritmo_significativo": {
            "type": "integer",
            "title": "Ejs Ritmo Significativo",
            "default": 0
          },
          "sde": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sde"
          },
          "crescimento_vs_ano_anterior": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crescimento Vs Ano Anterior"
          },
          "distribuicao_clusters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Distribuicao Clusters",
            "default": {

            }
          },
          "participacao_comunidades": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Participacao Comunidades",
            "default": {

            }
          }
        },
        "type": "object",
        "required": [
          "ano"
        ],
        "title": "IndicadoresRede",
        "description": "KPIs consolidados de toda a rede FEJEPE."
      },
...
      "RankingItem": {
        "properties": {
          "posicao": {
            "type": "integer",
            "title": "Posicao"
          },
          "id_ej": {
            "type": "integer",
            "title": "Id Ej"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "cluster": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster"
          },
          "comunidade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comunidade"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          },
          "foto_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto Url"
          }
        },
        "type": "object",
        "required": [
          "posicao",
          "id_ej",
          "nome",
          "valor"
        ],
        "title": "RankingItem",
        "description": "Item de um ranking de EJs."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}